+++ .vscode/launch.json
@@ -28,7 +28,19 @@
             "program": "${workspaceRoot}/src/azure-cli/azure/cli/__main__.py",
             "cwd": "${workspaceRoot}",
             "args": [
-                "--help",
+                "cdn",
+                "profile",
+                "cdn-migrate-to-afd",
+                "-g",
+                "cli-test-rg",
+                "--profile-name",
+                "cli-test-profile",
+                "--sku",
+                "Premium_AzureFrontDoor",
+                "--identity-type",
+                "SystemAssigned",
+                "--migration-endpoint-mappings",
+                "[{migrated-from:maxtestendpointcli-test-profile1.azureedge.net,migrated-to:maxtestendpointcli-test-profile2}]",
             ],
             "console": "externalTerminal",
             "debugOptions": [
