+++ src/azure-cli/azure/cli/command_modules/cdn/aaz/latest/cdn/profile_migration/_commit.py
@@ -12,16 +12,17 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "cdn profile cdn-migration-abort",
+    "cdn profile-migration commit",
+    is_preview=True,
 )
-class CdnMigrationAbort(AAZCommand):
-    """Abort the migration to Azure Frontdoor Premium/Standard.
+class Commit(AAZCommand):
+    """Commit the migrated Azure Frontdoor(Standard/Premium) profile.
     """
 
     _aaz_info = {
         "version": "2024-05-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/migrationabort", "2024-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/migrationcommit", "2024-05-01-preview"],
         ]
     }
 
@@ -29,7 +30,7 @@ class CdnMigrationAbort(AAZCommand):
 
     def _handler(self, command_args):
         super()._handler(command_args)
-        return self.build_lro_poller(self._execute_operations, self._output)
+        return self.build_lro_poller(self._execute_operations, None)
 
     _args_schema = None
 
@@ -44,14 +45,9 @@ class CdnMigrationAbort(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.profile_name = AAZStrArg(
             options=["--profile-name"],
-            help="Name of the Azure Front Door Standard or Azure Front Door Premium which is unique within the resource group.",
+            help="Name of the CDN profile which is unique within the resource group.",
             required=True,
             id_part="name",
-            fmt=AAZStrArgFormat(
-                pattern="^[a-zA-Z0-9]+(-*[a-zA-Z0-9])*$",
-                max_length=260,
-                min_length=1,
-            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -60,7 +56,7 @@ class CdnMigrationAbort(AAZCommand):
 
     def _execute_operations(self):
         self.pre_operations()
-        yield self.ProfilesMigrationAbort(ctx=self.ctx)()
+        yield self.ProfilesMigrationCommit(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -71,11 +67,7 @@ class CdnMigrationAbort(AAZCommand):
     def post_operations(self):
         pass
 
-    def _output(self, *args, **kwargs):
-        result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
-        return result
-
-    class ProfilesMigrationAbort(AAZHttpOperation):
+    class ProfilesMigrationCommit(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -105,7 +97,7 @@ class CdnMigrationAbort(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/migrationAbort",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/migrationCommit",
                 **self.url_parameters
             )
 
@@ -145,37 +137,12 @@ class CdnMigrationAbort(AAZCommand):
             }
             return parameters
 
-        @property
-        def header_parameters(self):
-            parameters = {
-                **self.serialize_header_param(
-                    "Accept", "application/json",
-                ),
-            }
-            return parameters
-
         def on_200(self, session):
-            data = self.deserialize_http_content(session)
-            self.ctx.set_var(
-                "instance",
-                data,
-                schema_builder=self._build_schema_on_200
-            )
-
-        _schema_on_200 = None
-
-        @classmethod
-        def _build_schema_on_200(cls):
-            if cls._schema_on_200 is not None:
-                return cls._schema_on_200
-
-            cls._schema_on_200 = AAZObjectType()
-
-            return cls._schema_on_200
+            pass
 
 
-class _CdnMigrationAbortHelper:
-    """Helper class for CdnMigrationAbort"""
+class _CommitHelper:
+    """Helper class for Commit"""
 
 
-__all__ = ["CdnMigrationAbort"]
+__all__ = ["Commit"]
