+++ src/azure-cli/azure/cli/command_modules/cdn/commands.py
@@ -107,7 +107,7 @@ def load_command_table(self, _):
     self.command_table['cdn endpoint rule remove'] = CDNEndpointRuleRemove(loader=self)
 
     from .custom.custom_cdn import CdnMigrateToAfd
-    self.command_table['cdn profile cdn-migrate-to-afd'] = CdnMigrateToAfd(loader=self)
+    self.command_table['cdn profile-migration migrate'] = CdnMigrateToAfd(loader=self)
 
     with self.command_group('cdn endpoint rule', cdn_endpoints_sdk, is_preview=True) as g:
         g.show_command('show', 'get')
